backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
qdrant-client
openai
anthropic
python-docx
pypdf
pydantic
PyJWT
